Further revise chunk implementation.
A new `PackedArray` type now covers previous use cases of `LongArray` and
`LightArray`. Two sets of bitwise reversals are applied on serialization
and deserialization, which has the effect of making contiguous all values.
The serialized representation is cached and dirtiness tracked.

NBT arrays now use packed array as their value type. The bits-per-block
is guessed based on the value length.

I'm not entirely convinced that this implementation is performant enough; I
may revise it to use `bytearray` or `numpy` or `array.array` while
retaining some of the ideas.